test: move public-API unit tests into the integration suite
Relocate inline `#[cfg(test)]` tests that only exercise the public API out of
`src/` and into `tests/`, so the source modules carry library code only:

- src/protocol.rs (10 tests)                   -> tests/protocol.rs
- src/miette_diagnostic.rs (2 tests)           -> tests/miette_diagnostic.rs
- src/eyreish/wrapper.rs (2 tests)             -> tests/wrapper.rs
- src/handlers/graphical/single_scan_tests.rs  -> tests/single_scan.rs

`crate::` paths become `miette::` and the per-test feature gates (`serde`,
`fancy`) are preserved. Tests that reach into `pub(crate)`/private internals
stay inline: `src/handlers/json.rs` (`escape`) and `src/source_impls.rs`
(`SpanScanner` and the line/column helpers).
